fix: skip new issue after updating

+++ dev/io.openliberty.data.internal_fat/test-applications/DataTestApp/src/test/jakarta/data/web/DataTestServlet.java
@@ -102,6 +102,9 @@ public class DataTestServlet extends FATServlet {
     Apartments apartments;
 
     @Inject
+    ApartmentsSpecific apartmentsSpecific;
+
+    @Inject
     EmptyRepository emptyRepo;
 
     @Inject
@@ -1115,6 +1118,10 @@ public class DataTestServlet extends FATServlet {
      */
     @Test
     public void testDeleteIgnoresFirstKeywork() {
+        if (skipForHibernateByDatabase("derby", "https://github.com/OpenLiberty/open-liberty/issues/33535")) {
+            return; // TODO remove once fixed in Hibernate - Cannot delete with pessimistic lock
+        }
+
         packages.deleteAll(); //cleanup before test
 
         packages.save(new Package(10001, 10.0f, 13.0f, 4.0f, "testDeleteIgnoresFirstKeywork#10001"));
@@ -1136,6 +1143,10 @@ public class DataTestServlet extends FATServlet {
      */
     @Test
     public void testDeleteQueryByParameters() {
+        if (skipForHibernateByDatabase("derby", "https://github.com/OpenLiberty/open-liberty/issues/33535")) {
+            return; // TODO remove once fixed in Hibernate - Cannot delete with pessimistic lock
+        }
+
         houses.dropAll();
 
         House h1 = new House();
@@ -1711,6 +1722,10 @@ public class DataTestServlet extends FATServlet {
      */
     @Test
     public void testFindAndDelete() {
+        if (skipForHibernateByDatabase("derby", "https://github.com/OpenLiberty/open-liberty/issues/33535")) {
+            return; // TODO remove once fixed in Hibernate - Cannot delete with pessimistic lock
+        }
+
         packages.save(new Package(40001, 41.0f, 14.0f, 4.0f, "testFindAndDelete#40001"));
         packages.save(new Package(40004, 44.0f, 40.4f, 4.4f, "testFindAndDelete#40004"));
         packages.save(new Package(40012, 42.0f, 12.0f, 2.0f, "testFindAndDelete#4001x"));
@@ -1771,6 +1786,10 @@ public class DataTestServlet extends FATServlet {
      */
     @Test
     public void testFindAndDeleteAnnotated() {
+        if (skipForHibernateByDatabase("derby", "https://github.com/OpenLiberty/open-liberty/issues/33535")) {
+            return; // TODO remove once fixed in Hibernate - Cannot delete with pessimistic lock
+        }
+
         packages.save(new Package(50001, 51.0f, 31.0f, 21.0f, "testFindAndDeleteAnnotated#50001"));
         packages.save(new Package(50002, 52.0f, 32.0f, 22.0f, "testFindAndDeleteAnnotated#50002"));
 
@@ -1801,6 +1820,10 @@ public class DataTestServlet extends FATServlet {
      */
     @Test
     public void testFindAndDeleteMultipleAnnotated() {
+        if (skipForHibernateByDatabase("derby", "https://github.com/OpenLiberty/open-liberty/issues/33535")) {
+            return; // TODO remove once fixed in Hibernate - Cannot delete with pessimistic lock
+        }
+
         packages.save(new Package(60001, 61.0f, 41.0f, 26.0f, "testFindAndDeleteMultipleAnnotated"));
         packages.save(new Package(60002, 62.0f, 42.0f, 25.0f, "testFindAndDeleteMultipleAnnotated"));
         packages.save(new Package(60003, 59.0f, 39.0f, 24.0f, "testFindAndDeleteMultipleAnnotated"));
@@ -1851,6 +1874,10 @@ public class DataTestServlet extends FATServlet {
         String jdbcJarName = System.getenv().getOrDefault("DB_DRIVER", "UNKNOWN");
         boolean supportsOrderByForUpdate = !jdbcJarName.startsWith("derby");
 
+        if (skipForHibernateByDatabase("derby", "https://github.com/OpenLiberty/open-liberty/issues/33535")) {
+            return; // TODO remove once fixed in Hibernate - Cannot delete with pessimistic lock
+        }
+
         packages.deleteAll();
 
         packages.save(new Package(80081, 18.0f, 18.1f, 8.8f, "testFindAndDeleteReturnsIds#80081"));
@@ -1890,6 +1917,10 @@ public class DataTestServlet extends FATServlet {
     public void testFindAndDeleteReturnsObjects() {
         String jdbcJarName = System.getenv().getOrDefault("DB_DRIVER", "UNKNOWN");
         boolean supportsOrderByForUpdate = !jdbcJarName.startsWith("derby");
+
+        if (skipForHibernateByDatabase("derby", "https://github.com/OpenLiberty/open-liberty/issues/33535")) {
+            return; // TODO remove once fixed in Hibernate - Cannot delete with pessimistic lock
+        }
 
         packages.deleteAll();
 
@@ -3820,9 +3851,26 @@ public class DataTestServlet extends FATServlet {
 
         assertEquals(true, multi.remove(added.get(2)));
 
-        assertEquals(true, multi.deleteById(908070605l).isPresent());
-        assertEquals(true, multi.deleteById(807060504l).isPresent());
-        assertEquals(false, multi.deleteById(706050403l).isPresent());
+        if (skipForHibernateByDatabase("derby", "https://github.com/OpenLiberty/open-liberty/issues/33535")) {
+            // TODO remove once fixed in Hibernate - Cannot delete with pessimistic lock
+            assertEquals(true, multi.remove(p1));
+            assertEquals(true, multi.remove(p2));
+            try {
+                assertEquals(false, multi.remove(p3));
+            } catch (OptimisticLockingFailureException e) {
+                //TODO we currently throw an exception when trying to delete an entity that does not exist.
+                //     however, shouldn't we just return false?
+            }
+        } else {
+            assertEquals(true, multi.deleteById(908070605l).isPresent());
+            assertEquals(true, multi.deleteById(807060504l).isPresent());
+            assertEquals(false, multi.deleteById(706050403l).isPresent());
+        }
+
+        if (skipForHibernateByDatabase("derby", "https://github.com/OpenLiberty/open-liberty/issues/33535")) {
+            packages.deleteAll();
+            return; // TODO remove once fixed in Hibernate - Cannot modify with pessimistic lock
+        }
 
         prod.name = "Test-Multiple-Entities-In-A-Repository-Product";
         assertEquals(1, multi.modify(prod));
@@ -4070,7 +4118,11 @@ public class DataTestServlet extends FATServlet {
      */
     @Test
     public void testPersistentFieldNamesAndDelimiters() {
-        apartments.removeAll();
+        if (skipForHibernateByDatabase("derby", "https://github.com/OpenLiberty/open-liberty/issues/33535")) {
+            apartmentsSpecific.removeAll(); // TODO remove once fixed in Hibernate - Cannot delete with pessimistic lock
+        } else {
+            apartments.removeAll();
+        }
 
         Apartment a101 = new Apartment();
         a101.occupant = new Occupant();
@@ -5984,6 +6036,10 @@ public class DataTestServlet extends FATServlet {
      */
     @Test
     public void testUpdateWithVersionedEntityParameter() {
+        if (skipForHibernateByDatabase("derby", "https://github.com/OpenLiberty/open-liberty/issues/33535")) {
+            return; // TODO remove once fixed in Hibernate - Cannot update with pessimistic lock
+        }
+
         Product prod1 = new Product();
         prod1.pk = UUID.nameUUIDFromBytes("UPD-VER-EP-1".getBytes());
         prod1.name = "testUpdateWithVersionedEntityParameter Product 1";
